Refactored dino datamapper class by introducing dino collection class
