+++ dino_catalog/dino_csv_data_mapper.rb
@@ -1,22 +1,20 @@
-# Map between dinosaur and CSV data source
 require 'csv'
-require 'json'
 require './dinosaur'
 
 class DinoCSVDataMapper
-  attr_accessor :dinosaurs
+  attr_reader :dinosaurs
 
-  def initialize
-    @data_source = ['dinodex.csv', 'african_dinosaur_export.csv']
+  def initialize(*data_source)
     @dinosaurs = []
+    data_source = ['dinodex.csv'] if data_source.empty?
 
-    @data_source.each do |ds|
+    data_source.each do |ds|
       raise "Data source: #{ds} not found" unless File.file? ds
-      load_data ds
+      import ds
     end
   end
 
-  def load_data(csv_file)
+  def import(csv_file)
     csv_data = CSV.read(csv_file, headers: true)
     csv_data.each do |csv_row|
 
@@ -48,38 +46,4 @@ class DinoCSVDataMapper
     end
   end
 
-  def find(condition = nil)
-    dinosaurs_found = DinoCSVDataMapper.new
-
-    return dinosaurs_found unless condition
-    raise "find condition must be a hash" unless condition.is_a? Hash
-
-    dinosaurs_found.dinosaurs = @dinosaurs
-
-    condition.map do |attr, val|
-      dinosaurs_found.dinosaurs = 
-        dinosaurs_found.dinosaurs.select do |dino_obj|
-          (ret = dino_obj.send(attr)) && ret.downcase == val.downcase
-        end
-    end
-
-    dinosaurs_found
-  end
-
-  TONS_TO_POUNDS = 2000
-  def find_large
-    dinosaurs_found = DinoCSVDataMapper.new
-    dinosaurs_found.dinosaurs = @dinosaurs
-    dinosaurs_found.dinosaurs.select do |dino_obj|
-      dino_obj.weight > 2 * TONS_TO_POUNDS
-    end
-  end
-
-  def to_s
-    @dinosaurs.inject('') { |string, dino| string << dino.to_s }
-  end
-
-  def to_json
-    @dinosaurs.map { |dino| dino.to_json }
-  end
 end
